Add project snapshot
