+++ JS/Projects.js
@@ -1,0 +1,18 @@
+//  Variable to store project objects and properties
+const workProperties = {
+  project1: {
+    featuredImage: '../images/Screenshot1.png',
+    name: 'To-Do List',
+    client: 'N/A',
+    role: 'Back End Dev',
+    year: '2015',
+    summary: '"To-do list" is a tool that helps to organize your day. It simply lists the things that you need to do and allows you to mark them as complete.',
+    description: 'A simple "To-do list" website that saves changes in the browser\'s local storage and uses ES6 and Webpack to add a new item, remove a selected item, mark a selected item as complete, remove all items marked as complete at once, and reorder a selected items (as drag-and-drop) from the list.',
+    technologies: ['html', 'css', 'javascript'],
+    liveVersionLink: 'https://c00p75.github.io/To-Do-List/dist',
+    sourceLink: 'https://github.com/c00p75/To-Do-List',
+    workHistoryDisplay: 'none',
+  },
+};
+
+export default workProperties;
