+++ JS/Script.js
@@ -1,3 +1,5 @@
+import workProperties from './Projects.js';
+
 //  Add click event listener to menu btn to toggle menu
 const expandedNav = document.querySelector('.menu-btn');
 const navLinks = document.querySelector('#desktop_nav');
@@ -21,23 +23,6 @@ navLinks.addEventListener('click', (event) => {
     mobileMenuClasses();
   }
 });
-
-//  Variable to store project objects and properties
-const workProperties = {
-  project1: {
-    featuredImage: '../images/Screenshot1.png',
-    name: 'To-Do List',
-    client: 'N/A',
-    role: 'Back End Dev',
-    year: '2015',
-    summary: 'A daily selection of privately personalized reads; no accounts or sign-ups required.',
-    description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent",
-    technologies: ['html', 'css', 'javascript'],
-    liveVersionLink: 'https://c00p75.github.io/To-Do-List/dist',
-    sourceLink: 'https://github.com/c00p75/To-Do-List',
-    workHistoryDisplay: 'none',
-  },
-};
 
 //  Function to addlist of technologies to portifolio project
 function addTechnologies(card) {
