+++ index.html
@@ -168,6 +168,6 @@
 				</div>
 			</section>
 	</main>
-	<script src="JS/Script.js"></script>
+	<script type="module" src="JS/Script.js"></script>
 </body>
 </html>
